Rename BaseSalesforceApiTask and BaseRetrieveMetadata
